requirements.txt
# requirements.txt
gradio==4.44.0
transformers==4.37.0
torch==2.1.0
torchaudio==2.1.0
twilio==8.10.0
openai-whisper==20231117
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
flask==2.3.3
python-dotenv==1.0.0
speechrecognition==3.10.0
pydub==0.25.1
google-cloud-speech==2.21.0
azure-cognitiveservices-speech==1.34.0
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
